test/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
psycopg[binary]==3.1.0
alembic==1.12.1
python-socketio==5.10.0
livekit==0.7.1
livekit-api==0.3.0
openai==1.3.0
deepgram-sdk==3.2.6
ffmpeg-python==0.2.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
pillow==10.1.0
python-jose[cryptography]==3.3.0

